@@ -1,13 +1,12 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery. DO NOT EDIT.
 
 package mocks
 
 import (
-	fs "io/fs"
+	"io/fs"
+	"time"
 
 	mock "github.com/stretchr/testify/mock"
-
-	time "time"
 )
 
 // FileInfo is an autogenerated mock type for the FileInfo type
@@ -86,28 +85,27 @@ func (_m *FileInfo) Size() int64 {
 }
 
 // Sys provides a mock function with given fields:
-func (_m *FileInfo) Sys() interface{} {
+func (_m *FileInfo) Sys() any {
 	ret := _m.Called()
 
-	var r0 interface{}
-	if rf, ok := ret.Get(0).(func() interface{}); ok {
+	var r0 any
+	if rf, ok := ret.Get(0).(func() any); ok {
 		r0 = rf()
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(interface{})
+			r0 = ret.Get(0).(any)
 		}
 	}
 
 	return r0
 }
 
-type mockConstructorTestingTNewFileInfo interface {
+// NewFileInfo creates a new instance of FileInfo. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewFileInfo(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewFileInfo creates a new instance of FileInfo. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewFileInfo(t mockConstructorTestingTNewFileInfo) *FileInfo {
+}) *FileInfo {
 	mock := &FileInfo{}
 	mock.Mock.Test(t)
 
