@@ -1,4 +1,4 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery. DO NOT EDIT.
 
 package mocks
 
@@ -39,13 +39,12 @@ func (_m *Check) Paths() []string {
 	return r0
 }
 
-type mockConstructorTestingTNewCheck interface {
+// NewCheck creates a new instance of Check. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewCheck(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewCheck creates a new instance of Check. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewCheck(t mockConstructorTestingTNewCheck) *Check {
+}) *Check {
 	mock := &Check{}
 	mock.Mock.Test(t)
 
