@@ -1,4 +1,4 @@
-// Code generated by mockery v2.14.0. DO NOT EDIT.
+// Code generated by mockery. DO NOT EDIT.
 
 package mocks
 
@@ -23,20 +23,23 @@ func (_m *File) Close() error {
 	return r0
 }
 
-// Seek provides a mock function with given fields: _a0, _a1
-func (_m *File) Seek(_a0 int64, _a1 int) (int64, error) {
-	ret := _m.Called(_a0, _a1)
+// Seek provides a mock function with given fields: offset, whence
+func (_m *File) Seek(offset int64, whence int) (int64, error) {
+	ret := _m.Called(offset, whence)
 
 	var r0 int64
+	var r1 error
+	if rf, ok := ret.Get(0).(func(int64, int) (int64, error)); ok {
+		return rf(offset, whence)
+	}
 	if rf, ok := ret.Get(0).(func(int64, int) int64); ok {
-		r0 = rf(_a0, _a1)
+		r0 = rf(offset, whence)
 	} else {
 		r0 = ret.Get(0).(int64)
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func(int64, int) error); ok {
-		r1 = rf(_a0, _a1)
+		r1 = rf(offset, whence)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -58,20 +61,23 @@ func (_m *File) Sync() error {
 	return r0
 }
 
-// Write provides a mock function with given fields: _a0
-func (_m *File) Write(_a0 []byte) (int, error) {
-	ret := _m.Called(_a0)
+// Write provides a mock function with given fields: p
+func (_m *File) Write(p []byte) (int, error) {
+	ret := _m.Called(p)
 
 	var r0 int
+	var r1 error
+	if rf, ok := ret.Get(0).(func([]byte) (int, error)); ok {
+		return rf(p)
+	}
 	if rf, ok := ret.Get(0).(func([]byte) int); ok {
-		r0 = rf(_a0)
+		r0 = rf(p)
 	} else {
 		r0 = ret.Get(0).(int)
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func([]byte) error); ok {
-		r1 = rf(_a0)
+		r1 = rf(p)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -79,34 +85,36 @@ func (_m *File) Write(_a0 []byte) (int, error) {
 	return r0, r1
 }
 
-// WriteAt provides a mock function with given fields: _a0, _a1
-func (_m *File) WriteAt(_a0 []byte, _a1 int64) (int, error) {
-	ret := _m.Called(_a0, _a1)
+// WriteAt provides a mock function with given fields: p, off
+func (_m *File) WriteAt(p []byte, off int64) (int, error) {
+	ret := _m.Called(p, off)
 
 	var r0 int
+	var r1 error
+	if rf, ok := ret.Get(0).(func([]byte, int64) (int, error)); ok {
+		return rf(p, off)
+	}
 	if rf, ok := ret.Get(0).(func([]byte, int64) int); ok {
-		r0 = rf(_a0, _a1)
+		r0 = rf(p, off)
 	} else {
 		r0 = ret.Get(0).(int)
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func([]byte, int64) error); ok {
-		r1 = rf(_a0, _a1)
+		r1 = rf(p, off)
 	} else {
 		r1 = ret.Error(1)
 	}
 
 	return r0, r1
-}
-
-type mockConstructorTestingTNewFile interface {
-	mock.TestingT
-	Cleanup(func())
 }
 
 // NewFile creates a new instance of File. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewFile(t mockConstructorTestingTNewFile) *File {
+// The first argument is typically a *testing.T value.
+func NewFile(t interface {
+	mock.TestingT
+	Cleanup(func())
+}) *File {
 	mock := &File{}
 	mock.Mock.Test(t)
 
