@@ -1,4 +1,4 @@
-// Code generated by mockery v2.16.0. DO NOT EDIT.
+// Code generated by mockery. DO NOT EDIT.
 
 package mocks
 
@@ -9,13 +9,13 @@ type ProgressBar struct {
 	mock.Mock
 }
 
-// Add provides a mock function with given fields: _a0
-func (_m *ProgressBar) Add(_a0 int) error {
-	ret := _m.Called(_a0)
+// Add provides a mock function with given fields: num
+func (_m *ProgressBar) Add(num int) error {
+	ret := _m.Called(num)
 
 	var r0 error
 	if rf, ok := ret.Get(0).(func(int) error); ok {
-		r0 = rf(_a0)
+		r0 = rf(num)
 	} else {
 		r0 = ret.Error(0)
 	}
@@ -37,20 +37,23 @@ func (_m *ProgressBar) Finish() error {
 	return r0
 }
 
-// Write provides a mock function with given fields: _a0
-func (_m *ProgressBar) Write(_a0 []byte) (int, error) {
-	ret := _m.Called(_a0)
+// Write provides a mock function with given fields: p
+func (_m *ProgressBar) Write(p []byte) (int, error) {
+	ret := _m.Called(p)
 
 	var r0 int
+	var r1 error
+	if rf, ok := ret.Get(0).(func([]byte) (int, error)); ok {
+		return rf(p)
+	}
 	if rf, ok := ret.Get(0).(func([]byte) int); ok {
-		r0 = rf(_a0)
+		r0 = rf(p)
 	} else {
 		r0 = ret.Get(0).(int)
 	}
 
-	var r1 error
 	if rf, ok := ret.Get(1).(func([]byte) error); ok {
-		r1 = rf(_a0)
+		r1 = rf(p)
 	} else {
 		r1 = ret.Error(1)
 	}
@@ -58,13 +61,12 @@ func (_m *ProgressBar) Write(_a0 []byte) (int, error) {
 	return r0, r1
 }
 
-type mockConstructorTestingTNewProgressBar interface {
+// NewProgressBar creates a new instance of ProgressBar. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
+// The first argument is typically a *testing.T value.
+func NewProgressBar(t interface {
 	mock.TestingT
 	Cleanup(func())
-}
-
-// NewProgressBar creates a new instance of ProgressBar. It also registers a testing interface on the mock and a cleanup function to assert the mocks expectations.
-func NewProgressBar(t mockConstructorTestingTNewProgressBar) *ProgressBar {
+}) *ProgressBar {
 	mock := &ProgressBar{}
 	mock.Mock.Test(t)
 
